test(skill-output): split validate_envelope tests; add reverse-gap coverage
tests/test_skill_output.py crossed the 500-line taste-lint gate again
after adding test_rejects_empty_message. Split TestValidateEnvelope
(the validator's own field-by-field tests, no dependency on the
producer-side functions) into a new tests/test_validate_envelope.py,
the same split pattern already used for the CLI subprocess tests.
test_skill_output.py now covers only get_output_format,
write_skill_output, and write_skill_error.

Added tests/test_skill_output_schema.py cases and
tests/test_skill_output.py::test_rejects_empty_message for the two
fixes in the sibling commit (write_skill_error's message guard,
Metadata.Script/Timestamp minLength), each proven to discriminate by
reverting the corresponding fix and confirming the exact predicted
failure.

uv run pytest tests/test_skill_output.py tests/test_skill_output_cli.py
tests/test_skill_output_schema.py tests/test_validate_envelope.py -q
-> 69 passed.

Refs #5201
